Organize common into race module
