--- simulation/common/__init__.py
@@ -1,5 +1,0 @@
-from simulation.common.race import (
-    Race,
-)
-
-Coordinate = tuple[float, float]
