--- simulation/common/exceptions.py
@@ -1,11 +1,0 @@
-"""
-Contains all the custom exceptions used in the simulation package
-"""
-
-
-class BatteryEmptyError(Exception):
-    pass
-
-
-class PrematureDataRecoveryError(Exception):
-    pass
