+++ simulation/race/__init__.py
@@ -1,0 +1,32 @@
+"""
+The `race` module contains utilities and classes relating to the
+manipulation of time-series data and localized data along a competition.
+"""
+
+from ._race import (
+    Race
+)
+
+from ._helpers import (
+    reshape_speed_array,
+    adjust_timestamps_to_local_times,
+    get_array_directional_wind_speed,
+    get_granularity_reduced_boolean,
+    calculate_completion_index,
+    get_map_data_indices,
+    parse_coordinates_from_kml
+)
+
+Coordinate = tuple[float, float]
+
+__all__ = [
+    "Race",
+    "reshape_speed_array",
+    "adjust_timestamps_to_local_times",
+    "get_array_directional_wind_speed",
+    "get_granularity_reduced_boolean",
+    "calculate_completion_index",
+    "get_map_data_indices",
+    "parse_coordinates_from_kml",
+    "Coordinate"
+]
